Added HomescreenViewModel
