+++ Quiz/Extensions/Extensions.swift
@@ -1,0 +1,15 @@
+//
+//  Extensions.swift
+//  Quiz
+//
+//  Created by Sandro Boka on 09.07.2024..
+//
+
+import Foundation
+import SwiftUI
+
+extension Color {
+    static let headerColor = Color(red: 64/255, green: 165/255, blue: 120/255)
+    static let backgroundColor = Color(red:255/255, green: 255/255, blue: 221/255)
+    static let buttonsColor = Color(red:21/255, green: 205/255, blue: 150/255)
+}
